Performance improvement: avoid conditional branches
Before this change, HTML::Escape used the obvious logic: it processes the
input one byte at a time, doing a series of comparisons on each one. This is
already much faster than doing a global regex replace in Perl. However, it
requires a large number of conditional branches for each input byte, and
in particular, a safe byte is subjected to N+1 conditional branches (where N
is the number of characters that get escaped, currently eight).

A faster approach avoids most of those conditionals: if a compiled-in array
is used to detect "unsafe" characters, the CPU can avoid doing those branches
except when needed. This optimises for strings with relatively few characters
that need to be escaped, but there's a significant improvement even on the
string in author/benchmark.pl, where 25% of the input bytes must be escaped.
My machine yielded these benchmark results under the previous version:

                 Rate HTML::Entities   HTML::Escape
HTML::Entities 26.3/s             --           -77%
HTML::Escape    112/s           327%             --

and these benchmark results with this change:

                 Rate HTML::Entities   HTML::Escape
HTML::Entities 25.1/s             --           -85%
HTML::Escape    168/s           568%             --